Update XLClass Break into separate files.
