+++ xlclass/globals.py
@@ -1,0 +1,8 @@
+from openpyxl.styles import PatternFill
+
+# Color dict for background fill
+COLORS = {'red': PatternFill(fgColor='FF0000', fill_type='solid'),
+          'green': PatternFill(fgColor='00b050', fill_type='solid'),
+          'orange': PatternFill(fgColor='FFC000', fill_type='solid'),
+          'yellow': PatternFill(fgColor='FFFF00', fill_type='solid'),
+          'gray': PatternFill(fgColor='C0C0C0', fill_type='solid')}
